manga/pyproject.toml
[project]
name = "manga"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "lxml>=6.0.1",
    "pillow>=11.3.0",
]

[project.scripts]
double-page-crop = "double_page_crop:main"
page-blender = "page_blender:main"
page-concat = "page_concat:main"
volume-split-into-chapters = "volume_split_into_chapters:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages]
find = { where = ["src"] }
